Desktop Applications/messaging_service/requirements.txt
adium-theme-ubuntu==0.3.4
cycler==0.9.0
matplotlib==1.5.1
numpy==1.11.0
Pillow==3.1.2
pycrypto==2.6.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2014.10
six==1.10.0
unity-lens-photos==1.0
virtualenv==15.0.1
